Add files via upload
A to-do list is a simple and effective way to keep track of tasks and responsibilities. It usually consists of a written or digital list where you can jot down all the tasks you need to complete. Each task is typically written as a short and concise description of what needs to be done.Its made using HTML , CSS AND JAVASCRIPT